Fix USB endpoint data subset.
On many systems, the USB took a long time to enumerate because we were replying with a 'nak' when we were finished sending a descriptor.  This is wrong.  When a descriptor or control transfer is complete, a DATAx packet should be sent of zero length.  This tells the USB subsystem "we're done"

Why doesn't anyone document this stuff well?